BugID: 200 Explicit use call an extra dpSet for the _stime attribute of the _online config. Also for BugId 219, BugId 223 and BugId 224 some services are added to the SAL sub-package.

+++ trunk/MAC/GCF/_PAL/PML/src/GCF_PropertySet.cc
@@ -76,7 +76,7 @@ GCFPropertySet::~GCFPropertySet()
 
 void GCFPropertySet::loadPropSetIntoRam()
 {
-  unsigned int sysNr = GCFPVSSInfo::getSysId(_scope);
+  int8 sysNr = GCFPVSSInfo::getSysId(_scope);
   GCFPVSSInfo::getTypeStruct(_type, _propSetInfo, sysNr);
   GCFProperty* pProperty;
   TPropertyInfo* pPropInfo;
